Don't use std::tuple as a poor man's compressed pair
…because it seems that MSVC doesn’t compress tuples like libstdc++ and
libc++.

Instead, use EBO by (indirectly) privately inheriting from the
Allocator — or rather by privately inheriting from a class which
privately inherits from the Allocator. The extra layer of indirection
is to make sure we can’t accidentally call this->allocate() or
this->construct() or what-have-you and end up calling the allocator
directly.