Remove is_default_constructible<value_type> test in default constructor
In the PIMPL test, MSVC doesn't like having the
is_default_constructible<V> test on the default constructor, complaining
that the type trait cannot be applied to incomplete types.

I'm not sure what's going on here, because by the time the compiler gets
as far as instantiating the constructor, the type should be complete. It
works fine with GCC and Clang. *shrug*